+++ scripts/ci/precheck_header.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-set -ex
+set -exv
 
 pip install -e ./tools
 
@@ -13,7 +13,7 @@ azdev verify document-map
 echo "Verify readme history"
 python -m automation.tests.verify_readme_history
 
-# # Only verify package version or PRs to Azure/azure-cli (not other forks)
+# Only verify package version or PRs to Azure/azure-cli (not other forks)
 # if [ $TRAVIS_EVENT_TYPE == "pull_request" ] && [ $TRAVIS_REPO_SLUG == "Azure/azure-cli" ]; then
 #     echo "Verify package versions"
 #    latestCliReleaseTag=$(git describe --tags `git rev-list --tags='azure-cli-*' --max-count=1`)