+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_help.py
@@ -349,6 +349,7 @@ helps['vmss start'] = """
 helps['vmss stop'] = """
     type: command
     short-summary: Power off (stop) VMs within a VMSS.
+    long-summary: The VMs will continue to be billed. To avoid this, you can deallocate VM instances within a VMSS through "az vmss deallocate"
 """
 
 helps['vmss update'] = """
@@ -545,28 +546,61 @@ helps['vm diagnostics set'] = """
     short-summary: Configure the Azure VM diagnostics extension.
     examples:
         - name: Set up default diagnostics on a Linux VM for Azure Portal VM metrics graphs and syslog collection.
-          text: >
-            # Set the following 3 parameters first.\n
-            my_resource_group=<Resource group name containing your Linux VM and the storage account>\n
-            my_linux_vm=<Your Azure Linux VM name>\n
-            my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>\n
+          text: |
+                # Set the following 3 parameters first.
+                my_resource_group=<Resource group name containing your Linux VM and the storage account>
+                my_linux_vm=<Your Azure Linux VM name>
+                my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>
 
-            my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_linux_vm --query "id" -o tsv)\n
+                my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_linux_vm --query "id" -o tsv)
 
-            default_config=$(az vm diagnostics get-default-config \\
-                | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
-                | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
+                default_config=$(az vm diagnostics get-default-config \\
+                    | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
+                    | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
 
-            storage_sastoken=$(az storage account generate-sas \\
-                --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
-                --permissions wlacu --resource-types co --services bt -o tsv)
+                storage_sastoken=$(az storage account generate-sas \\
+                    --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
+                    --permissions wlacu --resource-types co --services bt -o tsv)
 
-            protected_settings="{'storageAccountName': '{my_diagnostic_storage_account}', \\
-                'storageAccountSasToken': '{storage_sastoken}'}"
+                protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                    'storageAccountSasToken': '$storage_sastoken'}"
 
-            az vm diagnostics set --settings "{default_config}" \\
-                --protected-settings "{protected_settings}" \\
-                --resource-group $my_resource_group --vm-name $my_linux_vm
+                az vm diagnostics set --settings "$default_config" \\
+                    --protected-settings "$protected_settings" \\
+                    --resource-group $my_resource_group --vm-name $my_linux_vm
+
+        - name: Set up default diagnostics on a Windows VM.
+          text: |
+                # Set the following 3 parameters first.
+                my_resource_group=<Resource group name containing your Windows VM and the storage account>
+                my_windows_vm=<Your Azure Windows VM name>
+                my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>
+
+                my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_windows_vm --query "id" -o tsv)
+
+                default_config=$(az vm diagnostics get-default-config  --is-windows-os \\
+                    | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
+                    | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
+
+                # Please use the same options, the WAD diagnostic extension has strict
+                # expectations of the sas token's format. Set the expiry as desired.
+                storage_sastoken=$(az storage account generate-sas \\
+                    --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
+                    --permissions acuw --resource-types co --services bt --https-only --output tsv)
+
+                protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                    'storageAccountSasToken': '$storage_sastoken'}"
+
+                # # Alternatively, if the WAD extension has issues parsing the sas token,
+                # # one can use a storage account key.
+                # storage_account_key=$(az storage account keys list --account-name tosinstorage1win \\
+                #   --query [0].value -o tsv)
+                # protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                #   'storageAccountKey': '$storage_account_key'}"
+
+                az vm diagnostics set --settings "$default_config" \\
+                    --protected-settings "$protected_settings" \\
+                    --resource-group $my_resource_group --vm-name $my_windows_vm
 """
 
 disk_long_summary = """
@@ -969,6 +1003,17 @@ helps['vmss extension set'] = """
     - name: --name -n
       populator-commands:
       - az vm extension image list
+    examples:
+        - name: >
+                Set an extension which depends on two previously set extensions. That is, When a VMSS instance is
+                created or reimaged, the customScript extension will be provisioned only after all extensions that
+                it depends on have been provisioned. The extension need not depend on the other extensions for
+                pre-requisite configurations.
+          text: >
+                az vmss extension set --vmss-name my-vmss --name customScript --resource-group my-group \\
+                    --version 2.0 --publisher Microsoft.Azure.Extensions \\
+                    --provision-after-extensions NetworkWatcherAgentLinux VMAccessForLinux  \\
+                    --settings '{\"commandToExecute\": \"echo testing\"}'
 """
 
 helps['vmss extension show'] = """
@@ -1221,9 +1266,10 @@ helps['vm start'] = """
 
 helps['vm stop'] = """
     type: command
-    short-summary: Stop a running VM.
+    short-summary: Power off (stop) a running VM.
+    long-summary: The VM will continue to be billed. To avoid this, you can deallocate the VM through "az vm deallocate"
     examples:
-        - name: Stop a running VM.
+        - name: Power off (stop) a running VM.
           text: az vm stop -g MyResourceGroup -n MyVm
 {0}
 """.format(vm_ids_example.format('Stop all VMs in a resource group.', 'vm stop'))
@@ -1502,22 +1548,29 @@ helps['sig image-version create'] = """
     short-summary: creat a new image version
     long-summary : this operation might take a long time depending on the replicate region number. Use "--no-wait" is advised.
     examples:
-        - name: add a new image version
+        - name: Add a new image version
           text: |
             az sig image-version create -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --managed-image /subscriptions/00000000-0000-0000-0000-00000000xxxx/resourceGroups/imageGroups/providers/images/MyManagedImage
-        - name: add a new image version and don't wait on it. Later you can invoke "az sig image-version wait" command when ready to create a vm from the gallery image version
+        - name: Add a new image version replicated across multiple regions with different replication counts each. Eastus2 will have it's replica count set to the default replica count.
           text: |
-            az sig image-version create --no-wait -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --managed-image imageInTheSameResourceGroup
+                az sig image-version create -g MyResourceGroup --gallery-name MyGallery \\
+                --gallery-image-definition MyImage --gallery-image-version 1.0.0 \\
+                --managed-image image-name --target-regions eastus2 ukwest=3 southindia=2
+        - name: Add a new image version and don't wait on it. Later you can invoke "az sig image-version wait" command when ready to create a vm from the gallery image version
+          text: |
+            az sig image-version create --no-wait -g MyResourceGroup --gallery-name MyGallery \\
+            --gallery-image-definition MyImage --gallery-image-version 1.0.0 \\
+            --managed-image imageInTheSameResourceGroup
 """
 
 helps['sig image-version update'] = """
     type: command
     short-summary: update a share image version
     examples:
-        - name: replicate to a new set of regions
+        - name: Replicate to a new set of regions
           text: |
             az sig image-version update -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --target-regions westcentralus=2 eastus2
-        - name: replicate to one more region
+        - name: Replicate to one more region
           text: |
             az sig image-version update -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --add publishingProfile.targetRegions name=westcentralus