+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -3,15 +3,19 @@
 Release History
 ===============
 
+2.1.12
+++++++
+* Allow force push a helm chart that doesn't exist.
+
 2.1.11
 ++++++
 * Add support for image import from external Container Registries.
+* Condense the table layout for task list.
+* Add support for Azure DevOps URLs.
 
 2.1.10
 ++++++
 * Minor fixes.
-* Condense the table layout for task list.
-* Add support for Azure DevOps URLs.
 
 2.1.9
 +++++