+++ src/command_modules/azure-cli-find/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.2.13
+++++++
+* Minor fixes
+
 0.2.12
 ++++++
 * Minor fixes